Allow drawing shapegraphs (axial/segments)
Colour transfer is now done with PafColor instead of QRgb to avoid bringing QT into salalib. Added a 'built' bool in the various GL objects to check if the object has been created which will then allow it to be drawn (the app would crash in cases where there was no pointmap visible, i.e. when looking at an axial map)